feature: code-refactoring refactored text_validator start_query function
